refactor: extract initialiseFromQuery into its own file
